requirements.txt
gradio
pillow
tqdm
pillow-heif
imageio
numpy